feat(legacy-plugin-chart-event-flow): migrate package
