+++ packages/superset-ui-demo/package.json
@@ -29,6 +29,7 @@
   },
   "homepage": "https://github.com/apache-superset/superset-ui#readme",
   "dependencies": {
+    "@data-ui/event-flow": "^0.0.84",
     "@storybook/addon-actions": "^5.3.18",
     "@storybook/addon-info": "^5.3.18",
     "@storybook/addon-knobs": "^5.3.18",
@@ -42,6 +43,7 @@
     "@superset-ui/legacy-plugin-chart-calendar": "0.13.6",
     "@superset-ui/legacy-plugin-chart-chord": "0.13.13",
     "@superset-ui/legacy-plugin-chart-country-map": "0.13.3",
+    "@superset-ui/legacy-plugin-chart-event-flow": "^0.11.15",
     "@superset-ui/legacy-plugin-chart-force-directed": "0.13.16",
     "@superset-ui/legacy-plugin-chart-heatmap": "0.13.16",
     "@superset-ui/legacy-plugin-chart-histogram": "0.13.16",
