+++ plugins/legacy-plugin-chart-event-flow/src/controlPanel.jsx
