+++ plugins/legacy-plugin-chart-event-flow/src/images/thumbnail.png
