+++ plugins/legacy-plugin-chart-event-flow/src/images/thumbnailLarge.png
