--- temporary-plugins/superset-ui-plugins-demo/storybook/stories/legacy-plugin-chart-event-flow/index.js
@@ -1,8 +1,0 @@
-import EventFlowChartPlugin from '../../../../superset-ui-legacy-plugin-chart-event-flow';
-import Stories from './Stories';
-
-new EventFlowChartPlugin().configure({ key: 'event-flow' }).register();
-
-export default {
-  examples: [...Stories],
-};
